Simplify code and fix memory leak (#196)
* Simplify code and fix memory leak

* Delete obsolete file

* Pass functions as functions instead of creating anonymous functions
